refactor(modules): move audit log to general
