+++ src/i18n/Modules.js
@@ -514,4 +514,8 @@ export default {
     zh: '用户',
     en: 'User',
   },
+  auditLogDesc: {
+    zh: '审计日志功能可以记录对 EMQX 集群的关键操作，以满足企业在合规性和数据安全方面的需求。',
+    en: 'The audit log feature records critical operations on the EMQX cluster to meet enterprise requirements for compliance and data security.',
+  },
 }
